requirements.txt
requests==2.32.3
apache-airflow==3.0.1
pendulum==3.1.0
ruff>=0.11.6